GUI: fix warning
[JIRA] BTCUSTOMER-2402

+++ gui_engine/engine/acc_sw_rle_bypass.c
@@ -74,9 +74,9 @@ void rle_bypass_blit_2_rgb565(draw_img_t *image, struct gui_dispdev *dc,
         return;
     }
 
-    uint32_t image_off = sizeof(struct gui_rgb_data_head) + (uint32_t)(image->data);
+    uint32_t image_off = sizeof(struct gui_rgb_data_head) + (uint32_t)(uintptr_t)(image->data);
     uint8_t img_type = ((struct gui_rgb_data_head *)(image->data))->type;
-    imdc_file_t *file = (imdc_file_t *)image_off;
+    imdc_file_t *file = (imdc_file_t *)(uintptr_t)image_off;
     uint8_t opacity_value = image->opacity_value;
 
     if (opacity_value == 0)
@@ -107,10 +107,10 @@ void rle_bypass_blit_2_rgb565(draw_img_t *image, struct gui_dispdev *dc,
                 // memcpy line
                 uint8_t line_buf[BYTE_PIXEL_RGB565 * source_w];
                 uncompressed_rle_rgb565(file, line, line_buf);
-                int read_off = (int)line_buf - BYTE_PIXEL_RGB565 * image_x;
+                int read_off = (int)(uintptr_t)line_buf - BYTE_PIXEL_RGB565 * image_x;
                 for (uint32_t j = x_start; j <= x_end; j++)
                 {
-                    uint16_t pixel = *((uint16_t *)read_off + j);
+                    uint16_t pixel = *((uint16_t *)(uintptr_t)read_off + j);
                     if (pixel != 0)
                     {
                         writebuf[write_off + j] = ((((((pixel >> 11) << 3) * opacity_value + ((
